+++ src/notation/internal/notationactioncontroller.cpp
@@ -942,7 +942,7 @@ void NotationActionController::changeVoice(voice_idx_t voiceIndex)
     noteInput->setCurrentVoice(voiceIndex);
 
     if (!noteInput->isNoteInputMode()) {
-        interaction->changeSelectedNotesVoice(voiceIndex);
+        interaction->changeSelectedNotesVoice(static_cast<int>(voiceIndex));
     }
 }
